Changed back strcasecmp to int, bug in comparisons

+++ src/utils.h
@@ -30,6 +30,6 @@ int agon_fputs(char *s, uint8_t fileid);
 size_t agon_fwrite(void *ptr, size_t size, size_t nmemb, uint8_t fileid);
 size_t agon_fread(void *ptr, size_t size, size_t nmemb, uint8_t fileid);
 
-int8_t strcasecmp(char *s1, char *s2);
+int strcasecmp(char *s1, char *s2);
 
 #endif // UTILS_H
